♻️ Refactor: 체험 리스트 호출 코드 이동
